Replace catalog.json with manifest.json
As generating docs through `dbt docs generate` often takes a while, it is less economical to use `catalog.json` than `manifest.json` that gets created with every compile (`dbt compile` or as a part of `dbt run` or `dbt docs generate`). This could be done because we don't need information about the current state of database, only what's documented in dbt, because we can only create refs to these tables anyway.