Improve update of columns
When extracting columns from Superset, it was previously unclear which of the fields can cause errors when putting the columns back via PUT calls.
To remedy that, we chose a static list of fields to include when pushing the information back.

Now, Superset returns a clear error message, and it turns out it doesn't except "changed_on", "created_on", "type_generic" in any form
and "is_active" if it's empty. This is now reflected in the code.