Fix SQL parsing through `sqlfluff` (#25)
* Fix SQL parsing through `sqlfluff`

In its v1.3, `sqlfluff` [introduced breaking changes](https://docs.sqlfluff.com/en/stable/releasenotes.html#upgrading-to-1-3).
One of the things was a change to the parsing tree. To replace the original `sql_parsed.tree.recursive_crawl`, this introduces
own `crawl_dict_recursive` generator to apply on `sql_parsed`. On top of that, additional small changes were needed to
achieve the same output format and make the code easier to read and more performant.