Improve update of columns (#23)
When pushing columns to Superset, it was previously unclear which of the fields can cause errors in `PUT` calls. To remedy that, we chose a static list of fields to include when pushing the information back.

Now, Superset allows you to just pass `id` and `column_name` as unique column identifiers and whatever is not listed stays untouched. This is now reflected in the code whilst passing only `description` with it to be updated.